form fields restructure
