+++ package.json
@@ -20,7 +20,6 @@
     "eslint-config-airbnb-typescript": "^12.0.0",
     "eslint-plugin-jest": "^24.0.2",
     "firebase": "^7.19.1",
-    "prettier": "^2.1.2",
     "prop-types": "^15.7.2",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
